refactor: eliminate utilities
